Add files via upload
For clarification, this is a CONTINUATION of a previous mod that was discontinued. The dev didn't want to work on it anymore, and I am continuing to update this mod to keep it fresh and accurate for the community. The original code IS NOT MY OWN, but the updates which allow this mod to still work ARE.